Update dependencies and Dockerfile
This change updates the dependencies with `yarn upgrade`, so this should
respect the ranges set in `package.json`. Some dependencies were not in
the `yarn.lock` file, causing Cachito to not download all the dependencies
downstream and the downstream builds to fail.

This change also enables Dependabot to get automatic update proposals
and to keep the packages secure.

This change also changes the Dockerfile to use officially supported UBI
images provided by Red Hat. This allows us to use the same images as
downstream and thus to reduce risk of having different build results
between upstream and downstream. The `ubi8/nodejs-16` image ships
`yarn`, but it needs to be enable via `corepack`.

Signed-off-by: Fabien Dupont <fabiendupont@pm.me>